Rename components to YldLogoLink and ServiceLink to make them independant of TopNav component
