+++ src/components/Header/YldLogo.js
